requirements.txt
amqp==5.1.1
appdirs==1.4.3
apturl==0.5.2
asgiref==3.5.0
async-generator==1.10
attrs==19.3.0
Automat==20.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
billiard==3.6.4.0
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
cached-property==1.5.1
cachetools==5.1.0
cart==1.2.1
celery==5.2.6
certifi==2021.10.8
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.2.10
django-allauth==0.50.0
django-carton==1.2.1
django-ckeditor==6.4.1
django-cors-headers==3.11.0
django-environ==0.8.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-shopping-cart==0.1
djangorestframework==3.13.1
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.8.12.0
entrypoints==0.3
fake-useragent==0.1.11
fasteners==0.14.1
filelock==3.6.0
future==0.18.2
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
h11==0.13.0
httplib2==0.20.4
hyperlink==21.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==21.3.0
iniconfig==1.1.1
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.0
jsonschema==3.2.0
keyring==18.0.1
kombu==5.2.4
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpy==1.22.3
oauthlib==3.1.0
outcome==1.1.0
packaging==20.3
pandas==1.4.2
paramiko==2.6.0
parse==1.19.0
parsel==1.6.0
pexpect==4.6.0
Pillow==9.1.1
platformdirs==2.5.1
pluggy==1.0.0
prompt-toolkit==3.0.29
Protego==0.2.1
protobuf==3.20.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycryptodome==3.14.1
pycups==1.9.73
PyDispatcher==2.0.5
pydub==0.25.1
pyee==8.2.2
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==2.4.6
pyppeteer==1.0.2
pyquery==1.4.3
pyRFC3339==1.1
pyrsistent==0.15.5
PySocks==1.7.1
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.26
PyYAML==5.3.1
queuelib==1.6.2
razorpay==1.3.0
requests==2.22.0
requests-file==1.5.1
requests-html==0.10.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
rsa==4.8
Scrapy==2.6.1
SecretStorage==2.3.1
selenium==4.1.5
service-identity==21.1.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
ssh-import-id==5.10
systemd-python==234
texttable==1.6.2
tldextract==3.3.0
tomli==2.0.1
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
ttystatus==0.38
Twisted==22.4.0
typing-extensions==4.1.1
ubuntu-advantage-tools==27.7
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.9
usb-creator==0.3.7
vine==5.0.0
virtualenv==20.0.17
w3lib==1.22.0
wadllib==1.3.3
wcwidth==0.2.5
websocket-client==0.53.0
websockets==10.2
wget==3.2
wsproto==1.1.0
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
